+++ _BaseModule/changelog.md
@@ -1,4 +1,4 @@
-# v2.0.1r0 (02-NOV-25)
+# v2.0.2r0
 
-- Added Bascule Bridge sign replacement
-- Made 80km/h+ speed limit signs bigger
+- Now abbreviations are managed on WE window and stored inside the savegame, with default file for new cities.
+- Changing files location folder to the new Klyte45Mods path pattern